% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext.R
\docType{class}
\name{DelayedHDF5SMatrix-class}
\alias{DelayedHDF5SMatrix-class}
\alias{dim,H5S_dataset-method}
\alias{dim,DelayedHDF5SMatrix-method}
\alias{dimnames,DelayedHDF5SMatrix-method}
\alias{dimnames<-,DelayedHDF5SMatrix-method}
\alias{subset_seed_as_array,H5S_dataset-method}
\alias{[,DelayedHDF5SMatrix,ANY,ANY-method}
\alias{DelayedHDF5SMatrix}
\alias{show,DelayedHDF5SMatrix-method}
\title{define container for H5S_dataset to use Delayed* interfaces}
\usage{
\S4method{dim}{H5S_dataset}(x)

\S4method{dim}{DelayedHDF5SMatrix}(x)

\S4method{dimnames}{DelayedHDF5SMatrix}(x)

\S4method{dimnames}{DelayedHDF5SMatrix}(x) <- value

\S4method{subset_seed_as_array}{H5S_dataset}(seed, index)

\S4method{[}{DelayedHDF5SMatrix,ANY,ANY}(x, i, j, ..., drop = TRUE)

DelayedHDF5SMatrix(x, dimnames = list(NULL, NULL))

\S4method{show}{DelayedHDF5SMatrix}(object)
}
\arguments{
\item{x}{instance of \code{\link[rhdf5client]{H5S_dataset-class}}}

\item{value}{a list of dimnames}

\item{seed}{instance of H5S_dataset}

\item{index}{list}

\item{i}{selection vector or missing}

\item{j}{selection vector or missing}

\item{\dots}{not used}

\item{drop}{logical ignored for now}

\item{dimnames}{list or missing}

\item{object}{instance of \code{\link{DelayedHDF5SMatrix-class}}}
}
\description{
define container for H5S_dataset to use Delayed* interfaces

define a dim getter for H5S_dataset

define a dim getter for DelayedHDF5SMatrix

define dimnames operations for DelayedHDF5SMatrix

define dimnames setter for DelayedHDF5SMatrix

define subset_seed_as_array, key infrastructure for Delayed* seed

define bracket

constructor

printer
}
\section{Slots}{

\describe{
\item{\code{seed}}{instance of H5S_dataset}

\item{\code{dimnames}}{list}
}}

\examples{
library(rhdf5client)
if (!exists("bigec2")) example(H5S_source) # minimize hits to server
ds = bigec2[["assays"]] # methylation data
dem = DelayedHDF5SMatrix( ds )
dem
subset_seed_as_array(dem@seed, list(1:3, 1:4))
library(restfulSE)
data(banoSEMeta)
library(SummarizedExperiment)
assays(banoSEMeta)
assays(banoSEMeta)[[1]] = dem
assay(banoSEMeta[1:3,1:3])
assay(banoSEMeta[1:3,c("NA18498", "NA18501")])
}
